m3-sys: Plumb typename through import_global.
This is needed so const_int and const_UINT32 are not lowered,
and then m3c and C agree.

The point is to preserve const, to keep the "variables" in
read only memory, which is more efficient -- is unlikely
to suffer copy on write (unless it is mprotected), and
will remain backed by original executable not pagefile,
and can remain in CPU cached shared instead of exclusive.

Later better typing may also be possible.