m3core: Use WORD_T less. (#569)
Sometimes replace WORD_T with INTEGER, size_t, ssize_t, etc.
WORD_T is a confusing type because it does not exist in Modula-3.
Word.T is signed.

We cannot replace them all.

We can use the other types for "passthrough" when assignment (return,
parameter passing) is the only operation. We cannot clearly use
INTEGER or ssize_t in place of WORD_T when there is math.
We cannot clearly use size_t for parameters passed from Modula-3,
when i.e. they are INTEGER in Modula-3. The sizes would then
mismatch possibly on e.g. VMS.

For the M3WRAP stuff, size_t and ssize_t are good replacments
as they only do passthrough and they match closely .i3 lines.

We can use size_t or even int for "small indices" like
in the signal code.

For memory allocation, it in between.
On the Modula-3 side we have declared INTEGER.
On the C side we are calling functions that take size_t.
So there remains a problem there.