m3core, tcp: Remove Process_Exitor from C surface area.
Now that function pointers (ProcType) are untyped (char*),
we can no longer declare Process__RegisterExitor in C
and agree with m3c output.

As a reasonable workaround, return TRUE from IPInternal.Init (C)
and have Modula-3 then Process.RegisterExitor.

The Win32-specific exitor is then added to Unix, but the body
is internally ifdefed, so the function exists but does nothing,
and will never be called.