m3core: Remove vfork. (#647)
The declaration is incorrect for Solaris (sometimes int, sometimes long)
and causes build to fail, and the documentation is severely caveat laden,
and we don't use it.

Anyone using it can/should move a layer further into C and call vfork from C.
Or use fork instead. They are almost the same.