m3core: Add memcmp/memset/memmove/memcpy wrappers, for m3c to generate calls to. (#649)
The wrappers are just prefixed with "m3_".
Currently m3c declares and calls the C runtime functions directly.
This has been ok, and is a nice optimization, but a. it is extra
code in m3c b. it causes problems on Mac/arm64 (Mac/all?) when compiling as C++,
and I've always known it was dodgy, to declare these functions ourselves.