Reinstate _m3main in m3cc, as non-default option.
This sometimes greatly simplifies tracking down ways
m3gdb has been undermined by recent changes in libraries
and compilers.

Not all of the code for this was still present.  Had to find
it from old versions.

Retain main-in-C, as default.  New symbol M3_MAIN_M3CC makes it
happen, but only for internal or m3cc backend modes.  M3c is
not considered an internal mode, according to Target.i3.

There remain several occurrences of M3_MAIN_IN_C in config files,
but no code, before or after this change, that tests it.  These
are now all commented out in the config files.