m3c: When calling indirect, cast the the function pointer much more accurately.
Prior was cast to taking "() in C or "(...)" in C++.
That is, take anything.
( () in C++ means take nothing, (...) in C is not valid.)

This worked everywhere, except on Mac/arm64/C++, where it crashes.
Mac/arm64/C++ does not allow mixing "..." with strongly declared functions.
C++ works on all other platforms tried, a lot, and C works on Mac/arm64.
But specifically Mac/arm64/C++ did not work.

"more accurately" in that, now use cgtypes, not typeid or type.
This is still weakly typed but much better than before.
The weak proctype hashes continue to be in the way.
We also already use cgtype for the return.

Make call more expression based and less text based, as part of this,
since expressions -- each parameter -- can carry type information.

Reduce/remove used of CGType.Void. CGType.Addr is better even if weak.