m3core: Solaris: Type mismatch errors when m3c and .c files are concatenated. (#661)
In this case the mismatch is:

void RTLinker__AddUnit(RT0__Binder)

and then RT0__Binder is either correct like:
  struct RT0__Module;
  typedef RT0__Module* (__cdecl*RT0__Binder)(INTEGER mode);

or incorrect like:
  typedef void (__cdecl*M3PROC)(void);
  typedef M3PROC RT0__Binder;

We have to use the incorrect form because function pointers have many
type collision hashes. Even if it was not comment, because the return type
is ignored, the hashing is still weak and there is no provision for collisions.

We use M3PROC has shown because it at least matches in two places (atexit and RTProcess.RegisterExitor
and maybe others).

In time maybe m3c will ignore type hashes and will produce longer strings
that encode entire type structure with much less chance of collision.
Or at least use much better hash functions, like sha1 or better, of
such long strings. Or mostly require typenames. This structural type equivalence
was well motivated but is actually quite problematic.