Add "target64".
Add a limited form of 32bit targets that have
pointers and integers that are 64bits in size.

This provides two distinct advantages:
 - It cuts the number of bootstraps in half,
   if bootstrap is defined as building just cm3,
   and not the entire system.
   For example the X11 bindings are broken in this scheme.

 - It works around m3front regressions.
   Again, not for the entire system.

It provides "neutral":
 - It continues to work on 32bit hardware
   using 32bit instructions/registers, etc.

It provides not great:
 - Interoperability with native code is reduced.
   In particular, structs get the wrong layout.
   m3core is taught to handle this, with "#if M3_TARGET64'.

 - It doubles some costs.

 - The resulting minimal system is an ABI fork.

 - It is tempting to declare this "the 32bit system",
   but again, it breaks X11 and surely more.

It is not meant to be used with actual 64bit targets,
but should work with them with essentially no change.
Perhaps some code quality reduction (upleveling of all params and locals).

But again, the main point is to provide a
word-size agnostic .cpp file to build cm3.

The m3front regressions still need to be fixed.

It is not enough to just tell m3front that pointers or integers are 64bits.
They probably must be the same size.
And making them both 64bits is also not enough.
If you do that, what happens is we get native pointers, that m3front casts
address to a 64bit integer and writes, corrupting neighbor.

There are a few possible solutions to this.

One is:

templte <typename T>
struct M3Ptr
{
  union { T* p; INTEGER i; } data;
};

And change everything to this, including parameters, including in m3core/*.c.

This would likely work.

But another approach is tried here.
In this approach, m3core is changed less.
INTEGER is still widened, but pointers are not exactly, at least not as shown.

What we try here is:
 - upon function entry (begin_procedure) assign all parameters to locals;
   Er, well, make them all uplevel.
   This should be limited to pointers but is not.

 - Align all locals smaller than size 8 to 8.
   Er, well, append "[2]" to all uplevel declarations and "[0]" to all uplevel uses.
   That is, double all pointer sizes. Er, all locals. It should be limited to pointers.

 - C code needs no change to its parameters or locals.
   Except that INTEGER becomes long long. Pointers are still 32bits.
   Pointers in structs, of which are there are approximately zero in the base
   system, would need more attention.

Macros are now provided to provided "addresses" (ADDRESS, could be UINT64), to "pointer" (void*).
Because some compilers require this go through uintptr_t.

 This mode is triggered with cm3 -DM3_TARGET64
  cc -DM3_TARGET64
  scripts/*.py target64

64bit targets are not affected (well, you can use the target64 flag
and it does pessimize as described, and the results should work, but it isn't useful).

Non-C backends are not affected.
 Except there is a new mode of building m3core (i.e. C) that they
 cannot interoperate with.