mklib: Only export functions, not data. Modula-3 does not correctly access imported data. Also do not log differently bss vs. not.
This cuts AMD64_NT m3core exports from 4265 to 1528,
many of which I had added (WinConstants.c) and global
variables in ThreadWin32.

This is still more than really should be exported but it is progress.
i.e. We should export that which is defined in .i3 files,
not internals from .m3.