M3overrides cleanup.  - add some missing  - unify some old ones on the "new" way
For now at least, every m3makefile at the top of a package
should have an m3override file and every such m3override
should be identical, except for the very top one they all delegate
to.

On one hand this is kinda dumb, since it is all the same,
it should be automatic, but I think the idea is maybe to
accomodate multiple roots in a forest, not just one tree.

Probably that would be a reason to rename ROOT to CM3_ROOT,
so there can be other "ROOT"s.

This is an area that I don't think even other systems do great on.
There is a combination of
 - configure -with-foo=dir
 - moving under one root, one build like here (gcc+binutils+gmp, etc.)
 - install as you go, and find everything in common prefix,
   which is comman and seems ok until you really think about it,
   or see it in a setting where it fails 'cause
   1. I think the cross build methodology is make install DESTDIR=x and then
   later tar/untar, so there is no prefix to find during build and
   2. You end up running just-built things (if not cross building)
   which can be untested and fail.

For our part we have the one root approach, but we don't
practise much the deferred install or proper cross builds, etc.