OSF1: Switch Tru64 / OSF1 to C backend "only". (#724)
This is config and scripts.
The gcc backend did work a long time ago, and very recently also.
All of native cc, native cxx, gcc (3.4.6? 4.7?) and g++ should all work.
With native cxx, our install is slightly broken and linking requires -oldcxx,
just because one copy of one of the startup files is missing.
From a frontend point of view, old is likely Digital-authored
and "new" is EDG.

Fix path of C compiler in bootstrap.