scripts: Linking with -oldcxx breaks compiling Main on OSF. It has no long long, or no int64, or something. Probably neither. Workaround by having .o as dependency.
This depends on -o being restored.
And add the equivalent -Fo to the other NT platforms.
Add -o for OSF and Solaris.

This begs the question of portability to older compilers.

Really we should use autoconf and check sizef(long) == 8.
long long is not really needed on most 64bit targets.

Or maybe we should just require C99/stdint.h and be done
with a lot of cruft. We can still link with -oldcxx.