config: OSF: Suppress a few warnings.
C backend produces these:
 codeunreachable: statement is unreachable
 badansialias1: This statement uses the type to reference a storage location.
 badansialias2: This statement uses the type to reference the same storage location.
 fprldonot: IEEE dynamic rounding mode is not supported for long doubles on this platform
 labelnotreach: code can never be executed at label
 missingreturn: missing return statement at end of non-void function
 underflow: underflow occurs in evaluating this expression

Ideally we would work toward producing code that does not produce
any of these warnings, but it can be a fair amount of work,
for later/never.

Regarding underflow, we should generate IEEE754 bytes, not floats.
Including both endian and runtime selection.

Regarding rounding: We do not even use long doubles.
We get this warning because we are being even kind of advanced
and anticipating putting in the C99 support to change rounding mode,
which probably nobody calls.

missing return is probably in functions with infinite loops.

Augment comment about -oldcxx linking.