More adaptation sparked by 64-bit alignments on 32-bit targets.
Much of it response to the surprising (to me) discovery that
not all traced allocations have the same alignment.
Comments added to RT0.i3 summarize the confusing behavior.
After circumventing some apparently leftover stuff supporting
any of 8, 16, 32, or 64, all allocations end up aligned at
least to a target word, but on a 32-bit target, some can be
64-bit aligned.

Some comments and renames to save the next maintainer some
of the difficult vetting this required.

Also add some language section references to some compiler
messages.