add rpath support for Darwin platforms
This is to allow executables built using cm3 to find and use cm3
shared libraries.  rpath is now the preferred mechanism for locating
shared libraries on MacOS, but it is not 100% compatible with rpath on
other UN*Xes (b/c native ld does not support $ORIGIN, apparently), so
support is added to Darwin.common.