Fix cross-unit uses of formal default and field default array constructors.
Prior to the recent major rework of array constructors, there were
multiple copies of a constant array value, every time it was used.
Among others, two relevant ways multiple uses can occur are when the
constructor is anonymous and a default value of a parameter that is
used in more than one call or a default value of a field that is used
in more than one record construction or allocation.  The array rework
eliminated the duplicates, but requires some extra mechanism to access
the shared copy from a different unit.

This fix provides it by turning an anonymous array constructor in
either of these contexts into an artificial named constant.

This fixes p286, and two new tests p287 and p288.

Note that while the new tests check for correct values of array,
record, and set constructors, the latter two still do the duplication.