fix missing test report
Github seems to have a bug where test reports are not always associated with
their build.  As an attempt to work around this, I'm removing the reporting
workflow and adding the reports as a step in the build.  The expected
consequence is that we will no longer get test reports on pull requests, but
that's an acceptable tradeoff to not have to run around hunting for the reports
after a normal push.

More info at https://github.com/dorny/test-reporter/issues/67