add windows integration
I'm committing the windows integration even though it is broken.  Following
these steps by hand does work, but I've not yet been able to identify the
difference in the CI environment.

Mainly I want to have it present as part of the organization scheme for
separate fast and slow integration workflows I'm adding.