fix Windows integration test
Unfortunately limited in a couple of ways.

We can't "buildship all" because that runs out of memory in the Windows CI
environment, so we "buildship min" and the builds are not as comprehensive as
on the other platforms (linux/mac).

Also the m3tests suite does not work on Windows.  Each test is run as
"./pgm.exe" which Windows interprets as the command "." with the flag
"/pgm.exe", i.e., an error.  (This is reproducible outside of the CI
environment.)