Recode RecordExpr.Evaluate.  Fix 32-bit Windows bugs.
This problem has been here a long time, and had even been
viewed with suspicion.  Evaluate (formerly) replaced all
constant fields of a record constructor with their folded
values, in place, even if the whole record was not constant,
i.e., had a non-constant field.  But this undermined runtime
access to the constant fields when located in a different
interface, by stripping off the named constant in front of
them.

This change goes entirely functional by makeing a copy of
the RecordExpr.P and changing the args of the copy, leaving
the original unchanged.  The copy is made only if the whole
record is constant and a field is changed by folding.  Either
way, the copy or original is cached as the constant value of
the original.

This remains a bit dicey, in case of later changes to the fields
of the RecordExpr.P itself (not its 'args').

Similar rework of ArrayExpr.Evaluate would be needed, if ever
RecordExpr is fixed to avoid making copies of a static constructor
into every module that uses it, as ArrayExpr does now.

This problem existed in all targets, but was apparently seen
only on 32-bit Windows because of few instances of miscompiled
code, particularly in Codex86.m3:1067, which is only executed
when compiling for 32-bit Windows.