scripts: Remove the UpdateSource functionality.
This is an old unused disabled function the idea was
while porting, certain files were "likely" to be edited
and include them in the bootstrap.

It has long since been disabled.
The number of files edited in porting is now very much reduced.
They can be copied with scp, ssh, rsync, editing on target,
or git fetch, etc.

Granted, Git can be difficult to acquire, the copying
can be repitive and is nice to merge with copying of the bootstrap
archive, on-target editing can be difficult, and I have never used rsync.
I tend to rerun ssh/scp after each edit.