Scripts: Remove reload. (#800)
It does not work with python3 and appears from comments
not well motivated. Comments indicate an ability
to upgrade from NT386 to I386_NT or NT386 to I386_CYGWIN
or vice versa. This is better done with a rebootstrap these days,
which is not necessarily better or faster, but
is less hacky and must work anyway, vs. this little hack
for rare scenario.