m3core: Remove most checks of stack growth direction. (#812)
It is about as well to just check the direction of the given range.

Restore checking of register context, especially on grow-up stack.

Use alloca for stack pointer instead of address of local, to
better tolerate inlining. Platforms for which alloca is heap
instead of stack will "silently" not work -- they will likely
be very slow and/or free live data.

Put jmpbuf consistently in struct.

Gradually use unwrapped standard memset(0) idiom.