m3core: Adapt garbage collection to IA64 that has a second stack for register window flushing that grows up. Only HP-UX is implemented due to machine/documentation availbility. VMS, NT (Windows 7?), Linux, and maybe old FreeBSD would be nice.
Prior, we captured registers and flushed windows after sem_post
and before sigsuspend. The register capture seems unnecessary,
given we have a signal context, and the flush seems too late.
The capture is also outside of the stack.
In this change the capture is eliminated for non-SPARC/IA64
and the capture and flush is moved to before sem_post.

Turn off dynamic linking for HP-UX as the config is not correct for
no-gcc linker driver, and I am kinda tired of dynamic linking woes.
Though HP-UX does support $ORIGIN.