m3c: Use class and C-style casts. (#822)
Instead of typename and reinterpret_cast.
It seems both of these are needed for older OSF compiler
and that is ok. The template should maybe be removed.
C++ is most desirable, by far, for portable optimized
exception handling.
It *might* be desirable for other niceties like { }, C99
locals at not-top of scope before 99, math/fenv, bool, inline, etc.