m3c: Workaround m3front asking for types smaller than their size. m3front asks for pointers of size 8, on 32bit platforms, and then does write 8 bytes to them leading to variable overflows if we honor the type. For example in MD5.m3.
Possibly this generalizes to asking for pointers of arbitrary size
leading to arbitrary overflows?

This manifests at least sometimes as a warning from Visual C++.
  void* foo;
  * ( __int64* )&foo = 0; // warning C4739: reference to variable exceeds its storage space

Workaround: When type and size disagree, use the size.
That is, when given a non-struct with the wrong size, change it to a struct.
This a non-contradictory alternative to https://github.com/modula3/cm3/pull/829.

We fixup the IR in M3CCG_MultiPass in order achieve consistency in M3C.
There are no other users of M3CG_MultiPass, but all backends would
presumably prefer this behavior.

Fixing in m3front/CG might be even better, i.e. for LLVM.
But again, see https://github.com/modula3/cm3/pull/829 which
changes "stack type" of records from Addr to Struct, which
makes sense to me, but I am not an m3front expert.

Attempt to turn this warning into an error from Visual C++,
but the documented mechanism to do that does not work.
i.e. turning one warning into an error. Perhaps we should
use /WX to turn all warnings into errors but it is not
clearly a winnable battle.