Add m3-sys/m3middle/src/M3CG_Tee.[im]3.
This just splits operations on a M3CG.T into two other
child M3CG.Ts.  The great majority of this is just forwarding
each operation to each child.  For the few operations that
return values (Var, Proc, Label), it builds a map from each value
returned to the parent, to the pair received from the children.
For operations that receive such values, it uses the map to
forward correct values to the children.