djgpp: open and creat with binary mode. (#911)
Otherwise we get the size of a file, read it, and fail because
the read size and stat size vary.
I wonder if we should use stdio i.e. fopen("rb") instead of unistd.h.
Problem is still some need for opendir/readdir/closedir and stat,
which are not in stdio.h
Also fcntl would use file no. Maybe no/little gain.