DOS: Conditionally ignore mprotect of stack thread ends.
mprotect is built on DPMI 1.0.
DPMI 1.0 is rarely implemented.
DPMI 0.9 is more common.
When we mprotect the pages at the end of thread stacks,
this works under plain MS-DOS with presumably cwsdpmi.exe
but it fails under Windows (98SE). Ignore it for MS-DOS only.

Fix compile of user threads for not-DJGPP.