scripts: No longer delete config files in bin. (#951)
There appears to be a notion of deleting config files from the bin
directory (vs. the bin/config directory). I do not remember
where this came from, but I suspect it was to repair a brief
mistake long ago. Remove it.
Larger goal is to cleanup python scripts some and elevate
out of python directory.