scripts: Remove ShipBack from python. (#952)
I might come to regret this, but this is ongoing
removal of m3cc support from scripts\python,
esp. to trim scripts\python.
m3cc is supported by new concierge, and the .sh files,
and the scripts ultimately don't do much.
Three sets of scripts is excessive, for what
I do not use and does not support "a lot" of platforms.