scripts: Reduce the notion of installed config that reaches to source tree. (#954)
This was convenient to me long ago, to avoid copying over and over
but probably nobody else. Other automation is more heavily used now,
config files are more stable and simpler(fewer workarounds for older cm3),
it is not clearly as worthwhile. The reaching config remains if
someone wants to copy it manually. Hopefully people can keep
separate the installed vs. uninstalled tree.