m3core: Remove bitfields from RT0.
The goal of this PR is to remove endian-specific code from cm3
and its constituent libraries.

The reason for that, is so that the output of the C++ backend
needed to produce cm3, is portable between little endian and big endian systems.

The goal eventually is to have one thing to download for any system, Unix, Windows,
32bit, 64bit, little endian, big endian, just like most "source".

This is tedious and arguably slightly reduces readabilty and performance
but it should not be significant either way.