Endian neutrality: Trim libm3 for boot.
When building boot, remove from libm3: random ui sortedtable pickle hash.
These use bitfields and are not needed in cm3.

In m3c replace sortedtable with table to avoid use of bitfields.

Normal libm3 is not affected by this.
Only when building with cm3 -boot.

We should probably split into multiple libs but probably will not.